Calculate total number of textbooks per shipment
A college bookstore receives cartons of textbooks. In a shipment, each carton contains the same number of textbooks. The Inventory manager wants to use a computer to calculate the total number of textbooks arriving at the bookstore for each shipment, from the number of cartons and the number of textbooks in each carton. The inventory manager will enter the number of cartons received and the fixed number of textbooks in each carton for each shipment; the (WPF) application then will calculate the total number of textbooks in a shipment.